+++ setup.py
@@ -14,8 +14,8 @@ if ('CAIROCFFI_API_MODE' in os.environ and
 setup(
     name='cairocffi',
     use_scm_version=True,
-    install_requires=['cffi >= 1.1.0'],
-    setup_requires=['setuptools_scm', 'cffi >= 1.1.0'],
+    install_requires=['cffi >= 1.1.0', 'xcffib'],
+    setup_requires=['setuptools_scm', 'cffi >= 1.1.0', 'xcffib'],
     packages=['cairocffi'] if api_mode else [],
-    cffi_modules=['cairocffi/ffi.py:ffi']
+    cffi_modules=['cairocffi/ffi_build.py:ffi']
 )
